+++ modules/processing_vae.py
@@ -80,12 +80,15 @@ def full_vae_encode(image, model):
 
 
 def taesd_vae_decode(latents):
-    debug(f'VAE decode: name=TAESD images={len(latents)} latents={latents.shape}')
+    debug(f'VAE decode: name=TAESD images={len(latents)} latents={latents.shape} slicing={shared.opts.diffusers_vae_slicing}')
     if len(latents) == 0:
         return []
-    decoded = torch.zeros((len(latents), 3, latents.shape[2] * 8, latents.shape[3] * 8), dtype=devices.dtype_vae, device=devices.device)
-    for i in range(latents.shape[0]):
-        decoded[i] = sd_vae_taesd.decode(latents[i])
+    if shared.opts.diffusers_vae_slicing:
+        decoded = torch.zeros((len(latents), 3, latents.shape[2] * 8, latents.shape[3] * 8), dtype=devices.dtype_vae, device=devices.device)
+        for i in range(latents.shape[0]):
+            decoded[i] = sd_vae_taesd.decode(latents[i])
+    else:
+        decoded = sd_vae_taesd.decode(latents)
     return decoded